+++ ytdl_api/storage.py
@@ -3,8 +3,6 @@ import shutil
 from pathlib import Path
 from typing import Iterator
 
-from deta import Deta
-
 from .schemas.models import Download
 
 
@@ -64,31 +62,3 @@ class LocalFileStorage(IStorage):
     def remove_download_batch(self, storage_file_names: list[str]):
         for storage_file_name in storage_file_names:
             self.remove_download(storage_file_name)
-
-
-class DetaDriveStorage(IStorage):
-    """
-    Storage for saving downloaded media files in Deta Drive.
-    """
-
-    def __init__(self, deta_project_key: str, drive_name: str) -> None:
-        deta = Deta(project_key=deta_project_key)
-        self.drive = deta.Drive(drive_name)
-        # Amount of chunks read at a time by
-        self.chunk_size = 1024  # in bytes
-
-    def save_download_from_file(self, download: Download, path: Path) -> str:
-        self.drive.put(download.storage_filename, path=path)
-        return download.storage_filename
-
-    def get_download(self, storage_file_name: str) -> Iterator[bytes] | None:
-        file = self.drive.get(storage_file_name)
-        if file is None:
-            return file
-        return file.iter_chunks(self.chunk_size)
-
-    def remove_download(self, storage_file_name: str):
-        self.drive.delete(storage_file_name)
-
-    def remove_download_batch(self, storage_file_names: list[str]):
-        self.drive.delete_many(storage_file_names)